Ad exchange rate to t authorization banner
Also extract toThresholdTokenAmount function to utils
