+++ solidity-v1/dashboard/src/components/threshold/AuthorizeStakesBanner.jsx
@@ -4,6 +4,9 @@ import React from "react"
 import Chip from "../Chip"
 import { LINK } from "../../constants/constants"
 import OnlyIf from "../OnlyIf"
+import { KEEP, ThresholdToken } from "../../utils/token.utils"
+import BigNumber from "bignumber.js"
+import { toThresholdTokenAmount } from "../../utils/stake-to-t.utils"
 
 const AuthorizeStakesBanner = ({ numberOfStakesToAuthorize = 0 }) => {
   return (
@@ -38,7 +41,14 @@ const AuthorizeStakesBanner = ({ numberOfStakesToAuthorize = 0 }) => {
                     <Icons.Money
                       className={"authorize-stakes-banner__base-info-icon"}
                     />{" "}
-                    Exchange rate is 1 KEEP = 5 T.
+                    Exchange rate is 1 KEEP ={" "}
+                    {ThresholdToken.displayAmountWithSymbol(
+                      toThresholdTokenAmount(KEEP.fromTokenUnit(1)),
+                      3,
+                      (amount) =>
+                        new BigNumber(amount).toFormat(3, BigNumber.ROUND_DOWN)
+                    )}
+                    .
                   </li>
                 </ul>
               </div>
